寮食切り替え通知を外部関数化して、ON/OFFできるように(管理者限定コマンド)
